backend/requirements.txt
fastapi
uvicorn
authlib
python-dotenv
starlette
httpx
itsdangerous
redis
pymysql


########## ────────── FastAPI 서버 & 스트리밍 ──────────
# fastapi==0.115.12
# uvicorn==0.34.2              # --reload 로 개발 서버 실행
aiohttp==3.11.18             # 비동기 HTTP 클라이언트
aiofiles==24.1.0             # 파일 스트리밍 저장
# python-dotenv==1.1.0         # .env 로 OpenAI 키 로딩
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
# starlette==0.46.2            # FastAPI 의존

########## ────────── LangChain · OpenAI 챗봇 ──────────
openai==1.75.0
tiktoken==0.9.0
langchain==0.3.24
langchain-core==0.3.56
langchain-community==0.3.23
langchain-openai==0.3.15

########## ────────── 모델 추론 / 비전 파이프라인 ──────────
# PyTorch & 기본 비전 스택
# torch==2.3.0+cu118           # CU118 wheel 사용
# torchvision==0.18.0+cu118
timm==1.0.15                 # Swin-L 등 백본
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.2.1
pycocotools==2.0.8           # NMS & COCO util

# 빈 방 마스킹 + 인페인팅
diffusers==0.33.1
transformers==4.51.3
huggingface-hub==0.30.2
safetensors==0.5.3

# Grounded-SAM 파이프라인
# groundingdino-py==0.0.1      # PyPI 배포판 (또는 git 버전 주석 해제)
# Segment Anything
# git+https://github.com/facebookresearch/segment-anything.git


####### ────────── 데이터 · 보조 유틸 ──────────
requests==2.32.3
# httpx==0.28.1                # SSE 스트리밍
redis==6.0.0                 # 세션 메모리 캐시
SQLAlchemy==2.0.40           # (로그, 스키마 저장용 – 코드에 사용)
orjson==3.10.16              # 빠른 JSON 직렬화
typing_extensions==4.13.2
filelock==3.18.0
packaging==24.2
tqdm==4.67.1
scipy==1.15.2
scikit-learn==1.6.1          # 가구 개수 카운트 util
matplotlib==3.10.1           # (옵션) 모델 디버깅 / 결과 저장

